Rename lineage files to materialize.
Files were made before we agreed on name.
